Move installProjectTask to a helper file
